requirements.txt
streamlit
pandas
matplotlib
numpy
supabase
plotly
python-dotenv